+++ ak_library_management/wizard/borrow_books_checklist_wizard.py
@@ -3,6 +3,10 @@ from odoo import models, fields
 
 
 class BorrowBooksChecklistWizard(models.TransientModel):
+    """
+    this class used to create an object for raise popup at product template
+    message field contain warning or error message of the application.
+    """
     _name = "borrow.books.checklist.wizard"
     _description = "Borrow Books Checklist Wizard"
 
@@ -14,4 +18,4 @@ class BorrowBooksChecklistWizard(models.TransientModel):
         record deleted in the borrow_transaction_history table
         """
         rec = self.env.context.get('active_id')
-        self.env["borrow.transaction.history"].browse(rec).unlink()
+        self.env["borrow.transaction.history.wizard"].browse(rec).unlink()